Correccion de usuario controller y reposiitory. Creacion  de rutas controller y repository

+++ src/main/kotlin/com/es/proyectgoDinAda_GPS_Tracking/service/UsuarioService.kt
@@ -2,16 +2,19 @@ package com.es.proyectgoDinAda_GPS_Tracking.service
 
 import com.es.proyectgoDinAda_GPS_Tracking.exceptions.AlreadyExistsException
 import com.es.proyectgoDinAda_GPS_Tracking.exceptions.BadRequestException
+import com.es.proyectgoDinAda_GPS_Tracking.exceptions.ForbiddenException
 import com.es.proyectgoDinAda_GPS_Tracking.exceptions.NotFoundException
 import com.es.proyectgoDinAda_GPS_Tracking.model.Usuario
 import com.es.proyectgoDinAda_GPS_Tracking.repository.UsuarioRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.User
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
+import java.util.*
 
 @Service
 class UsuarioService: UserDetailsService {
@@ -23,36 +26,32 @@ class UsuarioService: UserDetailsService {
     private lateinit var usuarioRepository: UsuarioRepository
 
 
-    fun registerUsuario(usuario: Usuario): Usuario? {
+    fun registerUsuario(
+        usuario: Usuario
+    ): Usuario? {
+
+        checkEmptyFields(usuario)
+        checkPassword(usuario.password!!)
 
         // Comprobamos que el usuario (y email) no existe en la base de datos
-
         usuario.username?.let { usuarioRepository.findByUsername(it).ifPresent {throw AlreadyExistsException("El usuario ${usuario.username} ya existe")} }
 
         usuario.email?.let { usuarioRepository.findByEmail(it).ifPresent {throw AlreadyExistsException("El usuario con email ${usuario.email} ya existe")} }
 
 
-        // Creamos la instancia de Usuario
+        //Codificamos la contraseña, guarda el usuario en la BBDD y lo retorna
 
         usuario.password = passwordEncoder.encode(usuario.password)
 
         return usuarioRepository.save(usuario)
 
-
-        /*
-         La password del newUsuario debe estar hasheada, así que usamos el passwordEncoder que tenemos definido.
-         ¿De dónde viene ese passwordEncoder?
-         El objeto passwordEncoder está definido al principio de esta clase.
-         */
-        // Devolvemos el Usuario insertado en la BDD
     }
 
     /**
-     * Implementar la funcion de la interfaz
+     * Implementar la funcion de la interfaz. Esta funcion es usada por el autenticationmanager,  que us
      */
     override fun loadUserByUsername(username: String?): UserDetails {
-        val usuario: Usuario = usuarioRepository.findByUsername(username!!).orElseThrow {NotFoundException("El usuario no ha sido encontrado")}
-
+        val usuario: Usuario = findByUsername(username!!).orElseThrow {NotFoundException("El usuario no ha sido encontrado")}
 
         return User.builder()
             .username(usuario.username)
@@ -66,23 +65,35 @@ class UsuarioService: UserDetailsService {
         return usuarioRepository.findAll()
     }
 
-    fun getById(id: String): Usuario {
-        val usuario: Usuario = usuarioRepository.findByIdOrNull(id.toLong()) ?: throw NotFoundException("Usuario con ID $id no encontrado")
+
+    fun getById(
+        id: String,
+        authentication: Authentication
+    ): Usuario {
+
+        val usuario = findByIdOrThrow(id)
+
+        checkIfSelfOrAdmin(usuario, authentication)
 
         return usuario
     }
 
-    fun updateById(id: String, newUser: Usuario): Usuario{
-        val usuario = usuarioRepository.findByIdOrNull(id.toLong()) ?: throw NotFoundException("Usuario con id $id no ha sido encontrado")
 
-        checkEmptyFields(newUser)
-        checkNewName(usuario.username, newUser.username)
-        checkNewEmail(usuario.email, newUser.email)
+    fun updateById(
+        id: String,
+        newUser: Usuario,
+        authentication: Authentication
+    ): Usuario{
+
+        val usuario = findByIdOrThrow(id)
+
+        checkIfSelfOrAdmin(usuario, authentication)
+        validateNewUserData(newUser, usuario)
 
 
         // TODO: Hacer un check de estas opciones y otro de las nuevas que se ponen, para comprobar que no existan ya en la base de datos, en cuyo caso lanzar excepcion
         usuario.username = newUser.username ?: usuario.username
-        usuario.password = newUser.password ?: usuario.password
+        usuario.password = passwordEncoder.encode(newUser.password) ?: usuario.password
         usuario.roles = newUser.roles ?: usuario.roles
         usuario.email = newUser.email ?: usuario.email
 
@@ -90,13 +101,31 @@ class UsuarioService: UserDetailsService {
     }
 
 
-    fun deleteById(id: String) {
-        val usuario = usuarioRepository.findByIdOrNull(id.toLong()) ?: throw NotFoundException("Usuario con id $id no encontrado")
+    fun deleteById(
+        id: String,
+        authentication: Authentication
+    ) {
+        val usuario = findByIdOrThrow(id)
+
+        checkIfSelfOrAdmin(usuario, authentication)
 
         usuarioRepository.delete(usuario)
     }
 
 
+
+    /* VALIDACIONES Y COMPROBACIONES */
+
+    fun validateNewUserData(
+        newUser: Usuario,
+        usuario: Usuario
+    ){
+        checkEmptyFields(newUser)
+        checkNewName(usuario.username, newUser.username)
+        checkNewEmail(usuario.email, newUser.email)
+        // En este punto la password nunca sera nula ya que se com prueba antes
+        checkPassword(newUser.password!!)
+    }
 
     fun checkNewName(usuarioName: String?, newUserName: String?){
         if (newUserName != null && usuarioName != newUserName ){
@@ -111,9 +140,31 @@ class UsuarioService: UserDetailsService {
         }
     }
 
+    fun checkPassword(newPassword: String){
+        if (newPassword.length < 8) throw BadRequestException("La contraseña debe tener mas de 8 caracteres")
+    }
+
     fun checkEmptyFields(newUser: Usuario){
         if (newUser.email.isNullOrBlank()) throw BadRequestException("El email no puede estar vacío")
         if (newUser.username.isNullOrBlank()) throw BadRequestException("El nombre de usuario no puede estar vacío")
         if (newUser.password.isNullOrBlank()) throw BadRequestException("La contraseña no puede estar vacía")
+    }
+
+
+    fun checkIfSelfOrAdmin(user: Usuario, authentication: Authentication){
+        val roles = authentication.authorities.map { it.authority }
+
+        if (authentication.name != user.username && !roles.contains("ROLE_ADMIN")){
+            println("NOPE")
+            throw ForbiddenException("No tienes acceso a este recurso")
+        }
+    }
+
+    fun findByIdOrThrow(id: String): Usuario{
+        return usuarioRepository.findByIdOrNull(id.toLong()) ?: throw NotFoundException("Usuario con ID $id no encontrado")
+    }
+
+    fun findByUsername(username: String): Optional<Usuario> {
+        return usuarioRepository.findByUsername(username)
     }
 }
